Refactor: 패키지 정리, base entity 적용, 그외 기타 수정
