+++ src/main/java/com/project/devgram/DevgramApplication.java
@@ -2,8 +2,10 @@ package com.project.devgram;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class DevgramApplication {
 
 	public static void main(String[] args) {
