+++ src/main/java/com/project/devgram/controller/CommentServiceController.java
@@ -1,9 +1,8 @@
 package com.project.devgram.controller;
 
-import com.project.devgram.model.comment.CommentDto;
+import com.project.devgram.dto.CommentDto;
 import com.project.devgram.service.CommentService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/comments")
+@RequestMapping("api/comments")
 public class CommentServiceController {
 
     private final CommentService commentService;
@@ -20,9 +19,7 @@ public class CommentServiceController {
      * 댓글 등록 api
     */
     @PostMapping
-    public ResponseEntity<CommentDto> addComment(@RequestBody CommentDto commentInput) {
-        CommentDto commentDto = commentService.addComment(commentInput);
-
-        return ResponseEntity.ok(commentDto);
+    public CommentDto addComment(@RequestBody CommentDto commentInput) {
+        return commentService.addComment(commentInput);
     }
 }
