+++ src/test/java/com/project/devgram/service/CommentServiceTest.java
@@ -4,11 +4,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
+import com.project.devgram.dto.CommentDto;
 import com.project.devgram.entity.Comment;
-import com.project.devgram.model.comment.CommentDto;
-import com.project.devgram.model.comment.CommentStatus;
 import com.project.devgram.repository.CommentRepository;
-import java.time.LocalDateTime;
+import com.project.devgram.type.CommentStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -43,9 +42,10 @@ class CommentServiceTest {
             .parentCommentSeq(commentDto.getParentCommentSeq())
             .boardSeq(commentDto.getBoardSeq())
             .createdBy(commentDto.getCreatedBy())
-            .createdAt(LocalDateTime.now())
             .commentStatus(CommentStatus.POST)
             .build();
+
+        System.out.println(comment.getCommentSeq());
 
         given(commentRepository.save(any())).willReturn(comment);
